TINY-9277: Move dummy dialog into DialogUtils
